+++ app/Notifications/User/Register.php
@@ -32,7 +32,7 @@ class Register extends Notification implements ShouldQueue
   {
     return TelegramMessage::create()
       // Optional recipient user id.
-      ->to(env('TELEGRAM_CHAT_ID', '913493292'))
+      ->to((int)env('TELEGRAM_CHAT_ID', '913493292'))
       // Markdown supported.
       ->view('notification.user.registered', ['user' => $notifiable]);
   }
